requirements.txt
langchain
langchain-core

huggingface-hub
langchain-huggingface
python-dotenv
numpy
scikit-learn
streamlit